FastAPI/5_project1/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic 
python-jose[cryptography] 
passlib[bcrypt] 
python-dotenv 
pydantic-settings 
bcrypt